# client/utils/docs/openapi.yaml
openapi: "3.0.0"
info:
  title: "FAKE MESSAGE API"
  version: "1.0.0"
  description: |
    Fake Message 是在开发环境下发送邮件和短信的工具。因为在开发环境下，是不适合直接发送邮件和短信到真实的邮箱和手机的。该工具作为一个虚拟的邮件接收器，可模拟邮件和短信的接受。

    该文档列出发送邮件和短信的接口。通过调用接口的方式，新的以及历史的短信和邮件即可在该工具的列表页内展现。
  contact:
    name: "run"
    email: "run27017@126.com"
servers:
  - url: "http://localhost:3000/api"

paths:
  /emails/send:
    post:
      summary: "发送邮件"
      requestBody:
        description: "发送的邮件实体"
        content:
          application/json:
            schema:
              type: "object"
              properties:
                email:
                  type: "object"
                  properties:
                    toAddress:
                      type: "string"
                      description: "接受者的邮箱地址"
                      required: true
                    toName:
                      type: "string"
                      description: "接收者的姓名"
                    fromAddress:
                      type: "string"
                      description: "发送者的邮箱地址"
                      required: true
                    fromName:
                      type: "string"
                      description: "发送者的姓名"
                    tags:
                      type: "array"
                      description: "自定义的标签数组，标签对于标记邮件特别有用。"
                      items:
                        type: "string"
                    subject:
                      type: "string"
                      description: "邮件主题"
                      required: true
                    type:
                      type: "string"
                      description: "邮件正文的格式，可选值是`text`或`html`"
                      enum:
                        - "text"
                        - "html"
                    content:
                      type: "string"
                      description: "邮件正文的内容"
                      required: true
            example:
              email:
                toAddress: "receiver@example.com"
                toName: "接收者"
                fromAddress: "sender@example.com"
                fromName: "发送者"
                tags:
                  - "标签1"
                  - "标签2"
                subject: "邮件主题"
                type: "text"
                content: "邮件正文"
      responses:
        201:
          description: "返回发送成功的邮件"
          content:
            application/json:
              schema:
                properties:
                  email:
                    type: "object"
                    properties:
                      id:
                        type: "integer"
                        description: "新邮件的 id"
                      toAddress:
                        type: "string"
                        description: "接受者的邮箱地址"
                      toName:
                        type: "string"
                        description: "接收者的姓名"
                      fromAddress:
                        type: "string"
                        description: "发送者的邮箱地址"
                      fromName:
                        type: "string"
                        description: "发送者的姓名"
                      tags:
                        type: "array"
                        description: "自定义的标签数组，标签对于标记邮件特别有用。"
                        items:
                          type: "string"
                      subject:
                        type: "string"
                        description: "邮件主题"
                      type:
                        type: "string"
                        description: "邮件正文的格式，可选值是`text`或`html`"
                        enum:
                          - "text"
                          - "html"
                      content:
                        type: "string"
                        description: "邮件正文的内容"
                      createdAt:
                        type: "string"
                        format: "date-time"
                        description: "发送邮件的时间，格式始终是`YYYY-MM-DD hh:mm:ss`，时区以服务器所在时区为准。"
              example:
                email:
                  id: 1
                  toAddress: "receiver@example.com"
                  toName: "接收者"
                  fromAddress: "sender@example.com"
                  fromName: "发送者"
                  tags:
                    - "标签1"
                    - "标签2"
                  subject: "邮件主题"
                  type: "text"
                  content: "邮件正文"
                  createdAt: "2019-07-22 18:06:06"
  /emails/xsend:
    post:
      summary: "通过模板发送邮件"
      requestBody:
        description: "发送的邮件实体"
        content:
          application/json:
            schema:
              type: "object"
              properties:
                email:
                  type: "object"
                  properties:
                    toAddress:
                      type: "string"
                      description: "接受者的邮箱地址"
                      required: true
                    toName:
                      type: "string"
                      description: "接收者的姓名"
                    fromAddress:
                      type: "string"
                      description: "发送者的邮箱地址"
                      required: true
                    fromName:
                      type: "string"
                      description: "发送者的姓名"
                    tags:
                      type: "array"
                      description: "自定义的标签数组，标签对于标记邮件特别有用。"
                      items:
                        type: "string"
                    subject:
                      type: "string"
                      description: "邮件主题"
                      required: true
                    templateId:
                      type: "number"
                      description: "模板 id"
                      required: true
                    bindings:
                      type: "object"
                      description: "变量绑定的键值对"
            example:
              email:
                toAddress: "receiver@example.com"
                toName: "接收者"
                fromAddress: "sender@example.com"
                fromName: "发送者"
                tags:
                  - "标签1"
                  - "标签2"
                templateId: 1
                bindings:
                  NAME: '李先生'
      responses:
        201:
          description: "返回发送成功的邮件"
          content:
            application/json:
              schema:
                properties:
                  email:
                    type: "object"
                    properties:
                      id:
                        type: "integer"
                        description: "新邮件的 id"
                      toAddress:
                        type: "string"
                        description: "接受者的邮箱地址"
                      toName:
                        type: "string"
                        description: "接收者的姓名"
                      fromAddress:
                        type: "string"
                        description: "发送者的邮箱地址"
                      fromName:
                        type: "string"
                        description: "发送者的姓名"
                      tags:
                        type: "array"
                        description: "自定义的标签数组，标签对于标记邮件特别有用。"
                        items:
                          type: "string"
                      subject:
                        type: "string"
                        description: "邮件主题"
                      type:
                        type: "string"
                        description: "邮件正文的格式，可选值是`text`或`html`"
                        enum:
                          - "text"
                          - "html"
                      content:
                        type: "string"
                        description: "邮件正文的内容"
                      createdAt:
                        type: "string"
                        format: "date-time"
                        description: "发送邮件的时间，格式始终是`YYYY-MM-DD hh:mm:ss`，时区以服务器所在时区为准。"
              example:
                email:
                  id: 1
                  toAddress: "receiver@example.com"
                  toName: "接收者"
                  fromAddress: "sender@example.com"
                  fromName: "发送者"
                  tags:
                    - "标签1"
                    - "标签2"
                  subject: "邮件主题"
                  type: "text"
                  content: "邮件正文"
                  createdAt: "2019-07-22 18:06:06"
  /messages/send:
    post:
      summary: "发送短信"
      requestBody:
        description: "发送的短信实体"
        content:
          application/json:
            schema:
              properties:
                message:
                  type: "object"
                  properties:
                    toMobile:
                      type: "string"
                      description: "接受者的手机号"
                      required: true
                    tags:
                      type: "array"
                      description: "自定义的标签数组，标签对于标记短信特别有用。"
                      items:
                        type: "string"
                    content:
                      type: "string"
                      description: "短信正文的内容"
                      required: true
            example:
              message:
                toMobile: "#001"
                tags:
                  - "标签1"
                  - "标签2"
                content: "短信正文"
      responses:
        201:
          content:
            application/json:
              description: "返回发送成功的短信"
              schema:
                properties:
                  message:
                    type: "object"
                    properties:
                      id:
                        type: "integer"
                        description: "新短信的 id"
                      toMobile:
                        type: "string"
                        description: "接受者的手机号"
                      tags:
                        type: "array"
                        description: "自定义的标签数组，标签对于标记短信特别有用。"
                        items:
                          type: "string"
                      content:
                        type: "string"
                        description: "短信正文的内容"
                      createdAt:
                        type: "string"
                        format: "date-time"
                        description: "发送短信的时间，格式始终是`YYYY-MM-DD hh:mm:ss`，时区以服务器所在时区为准。"
              example:
                message:
                  id: 1
                  toMobile: "#001"
                  tags:
                    - "标签1"
                    - "标签2"
                  content: "短信正文"
                  createdAt: "2019-07-22 18:06:06"
  /messages/xsend:
    post:
      summary: "通过模板发送短信"
      requestBody:
        description: "发送的短信实体"
        content:
          application/json:
            schema:
              properties:
                message:
                  type: "object"
                  properties:
                    toMobile:
                      type: "string"
                      description: "接受者的手机号"
                      required: true
                    tags:
                      type: "array"
                      description: "自定义的标签数组，标签对于标记短信特别有用。"
                      items:
                        type: "string"
                    templateId:
                      type: "number"
                      description: "模板 id"
                      required: true
                    bindings:
                      type: "object"
                      description: "变量绑定的键值对"
            example:
              message:
                toMobile: "#001"
                tags:
                  - "标签1"
                  - "标签2"
                templateId: 2
                bindings:
                  CODE: '121212'
      responses:
        201:
          content:
            application/json:
              description: "返回发送成功的短信"
              schema:
                properties:
                  message:
                    type: "object"
                    properties:
                      id:
                        type: "integer"
                        description: "新短信的 id"
                      toMobile:
                        type: "string"
                        description: "接受者的手机号"
                      tags:
                        type: "array"
                        description: "自定义的标签数组，标签对于标记短信特别有用。"
                        items:
                          type: "string"
                      content:
                        type: "string"
                        description: "短信正文的内容"
                      createdAt:
                        type: "string"
                        format: "date-time"
                        description: "发送短信的时间，格式始终是`YYYY-MM-DD hh:mm:ss`，时区以服务器所在时区为准。"
              example:
                message:
                  id: 1
                  toMobile: "#001"
                  tags:
                    - "标签1"
                    - "标签2"
                  content: "短信正文"
                  createdAt: "2019-07-22 18:06:06"